# apps/backend/openapi.json
{
	"openapi": "3.0.3",
	"info": {
		"description": "API's used and provided by [https://StellarAtlas.io](https://StellarAtlas.io).",
		"title": "StellarAtlas.io API",
		"contact": {
			"name": "StellarAtlas",
			"email": "info@StellarAtlas.io"
		},
		"version": "v1"
	},
	"servers": [
		{
			"url": "https://api.StellarAtlas.io",
			"description": "API for Stellar public network"
		},
		{
			"url": "https://api-testnet.StellarAtlas.io",
			"description": "API for Stellar testnet"
		}
	],
...
	"paths": {
...
		"/v1/search": {
			"get": {
				"summary": "Search network entities",
				"description": "Faceted search for current validators, nodes, organizations, and archive-related network metadata.",
				"tags": ["Search", "Network", "Node", "Organization"],
				"operationId": "searchNetworkEntities",
				"parameters": [
					{
						"name": "q",
						"in": "query",
						"description": "Search text",
						"required": false,
						"schema": {
							"type": "string"
						}
					},
					{
						"name": "type",
						"in": "query",
						"description": "Optional entity type filter",
						"required": false,
						"schema": {
							"type": "string",
							"enum": ["node", "organization"]
						}
					},
					{
						"name": "active",
						"in": "query",
						"description": "Optional active entity filter",
						"required": false,
						"schema": { "type": "boolean" }
					},
					{
						"name": "archiveStatus",
						"in": "query",
						"description": "Optional node archive status filter",
						"required": false,
						"schema": {
							"type": "string",
							"enum": ["error", "ok", "unknown"]
						}
					},
					{
						"name": "countryCode",
						"in": "query",
						"description": "Optional node country-code filter",
						"required": false,
						"schema": { "type": "string" }
					},
					{
						"name": "organizationId",
						"in": "query",
						"description": "Optional organization id filter",
						"required": false,
						"schema": { "type": "string" }
					},
					{
						"name": "validator",
						"in": "query",
						"description": "Optional validator flag filter",
						"required": false,
						"schema": { "type": "boolean" }
					},
					{
						"name": "fullValidator",
						"in": "query",
						"description": "Optional full-validator flag filter",
						"required": false,
						"schema": { "type": "boolean" }
					},
					{
						"name": "validating",
						"in": "query",
						"description": "Optional currently-validating flag filter",
						"required": false,
						"schema": { "type": "boolean" }
					},
					{
						"name": "topTier",
						"in": "query",
						"description": "Optional top-tier flag filter",
						"required": false,
						"schema": { "type": "boolean" }
					},
					{
						"name": "limit",
						"in": "query",
						"description": "Maximum results to return",
						"required": false,
						"schema": {
							"type": "integer",
							"minimum": 1,
							"maximum": 25
						}
					}
				],
				"responses": {
					"200": {
						"description": "Search results",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/SearchResponse"
								}
							}
						}
					}
				}
			}
		},
...
		"/v1/explorer/search": {
			"get": {
				"summary": "Search the Horizon-backed explorer helpers",
				"description": "Searches transaction hashes, account IDs, ledger sequences, asset queries, and contract IDs using the currently configured Horizon/RPC helper paths.",
				"tags": ["Explorer", "Horizon"],
				"operationId": "searchExplorer",
				"parameters": [
					{
						"name": "query",
						"in": "query",
						"required": true,
						"schema": { "type": "string" }
					},
					{
						"name": "type",
						"in": "query",
						"required": false,
						"schema": {
							"type": "string",
							"enum": [
								"account",
								"asset",
								"auto",
								"contract",
								"ledger",
								"transaction"
							]
						}
					}
				],
				"responses": {
					"200": {
						"description": "Explorer search result",
						"content": {
							"application/json": {
								"schema": { "$ref": "#/components/schemas/ExplorerSearch" }
							}
						}
					},
					"400": { "description": "Invalid search input" },
					"502": { "description": "Explorer search unavailable" }
				}
			}
		},
		"/v1/explorer/transactions": {
			"get": {
				"summary": "Get recent transactions",
				"description": "Returns a bounded recent transaction feed from configured Horizon.",
				"tags": ["Explorer", "Horizon"],
				"operationId": "getExplorerRecentTransactions",
				"parameters": [
					{
						"name": "limit",
						"in": "query",
						"required": false,
						"schema": { "type": "integer", "minimum": 1, "maximum": 50 }
					}
				],
				"responses": {
					"200": {
						"description": "Recent transactions",
						"content": {
							"application/json": {
								"schema": { "$ref": "#/components/schemas/RecentTransactions" }
							}
						}
					},
					"400": { "description": "Invalid transaction limit" },
					"502": { "description": "Transaction feed unavailable" }
				}
			}
		},
		"/v1/explorer/transactions/{hash}": {
			"get": {
				"summary": "Look up an explorer transaction",
				"description": "Explorer namespace alias for Horizon-backed transaction lookup.",
				"tags": ["Explorer", "Horizon"],
				"operationId": "getExplorerTransactionByHash",
				"parameters": [
					{
						"name": "hash",
						"in": "path",
						"required": true,
						"schema": { "type": "string", "pattern": "^[0-9a-fA-F]{64}$" }
					}
				],
				"responses": {
					"200": {
						"description": "Transaction lookup result",
						"content": {
							"application/json": {
								"schema": { "$ref": "#/components/schemas/TransactionLookup" }
							}
						}
					},
					"400": { "description": "Invalid transaction hash" },
					"404": { "description": "Transaction not found" },
					"502": { "description": "Transaction lookup unavailable" }
				}
			}
		},
		"/v1/explorer/transactions/{hash}/operations": {
			"get": {
				"summary": "Get operations for a transaction",
				"description": "Returns bounded Horizon operation rows for one transaction hash.",
				"tags": ["Explorer", "Horizon"],
				"operationId": "getExplorerTransactionOperations",
				"parameters": [
					{
						"name": "hash",
						"in": "path",
						"required": true,
						"schema": { "type": "string", "pattern": "^[0-9a-fA-F]{64}$" }
					}
				],
				"responses": {
					"200": {
						"description": "Transaction operation rows",
						"content": {
							"application/json": {
								"schema": { "$ref": "#/components/schemas/ExplorerOperations" }
							}
						}
					},
					"400": { "description": "Invalid transaction hash" },
					"404": { "description": "Transaction not found" },
					"502": { "description": "Transaction operations unavailable" }
				}
			}
		},
		"/v1/explorer/ledgers/{sequence}": {
			"get": {
				"summary": "Look up a ledger",
				"description": "Returns ledger facts from configured Horizon.",
				"tags": ["Explorer", "Horizon"],
				"operationId": "getExplorerLedger",
				"parameters": [
					{
						"name": "sequence",
						"in": "path",
						"required": true,
						"schema": { "type": "string", "pattern": "^\\d+$" }
					}
				],
				"responses": {
					"200": {
						"description": "Ledger lookup result",
						"content": {
							"application/json": {
								"schema": { "$ref": "#/components/schemas/ExplorerLedger" }
							}
						}
					},
					"400": { "description": "Invalid ledger sequence" },
					"404": { "description": "Ledger not found" },
					"502": { "description": "Ledger lookup unavailable" }
				}
			}
		},
		"/v1/explorer/accounts/{accountId}": {
			"get": {
				"summary": "Look up an account",
				"description": "Returns account state from configured Horizon.",
				"tags": ["Explorer", "Horizon"],
				"operationId": "getExplorerAccount",
				"parameters": [
					{
						"name": "accountId",
						"in": "path",
						"required": true,
						"schema": { "type": "string" }
					}
				],
				"responses": {
					"200": {
						"description": "Account lookup result",
						"content": {
							"application/json": {
								"schema": { "$ref": "#/components/schemas/ExplorerAccount" }
							}
						}
					},
					"400": { "description": "Invalid account address" },
					"404": { "description": "Account not found" },
					"502": { "description": "Account lookup unavailable" }
				}
			}
		},
		"/v1/explorer/assets": {
			"get": {
				"summary": "Search assets",
				"description": "Returns a bounded Horizon asset search result by optional code and issuer.",
				"tags": ["Explorer", "Horizon"],
				"operationId": "getExplorerAssets",
				"parameters": [
					{
						"name": "assetCode",
						"in": "query",
						"required": false,
						"schema": { "type": "string" }
					},
					{
						"name": "assetIssuer",
						"in": "query",
						"required": false,
						"schema": { "type": "string" }
					}
				],
				"responses": {
					"200": {
						"description": "Asset search result",
						"content": {
							"application/json": {
								"schema": { "$ref": "#/components/schemas/ExplorerAssets" }
							}
						}
					},
					"400": { "description": "Invalid asset issuer" },
					"502": { "description": "Asset search unavailable" }
				}
			}
		},
		"/v1/explorer/operations": {
			"get": {
				"summary": "Search operations",
				"description": "Returns bounded Horizon operation rows filtered by ledger, account, operation type, or date range.",
				"tags": ["Explorer", "Horizon"],
				"operationId": "getExplorerOperations",
				"parameters": [
					{
						"name": "accountId",
						"in": "query",
						"required": false,
						"schema": { "type": "string" }
					},
					{
						"name": "ledger",
						"in": "query",
						"required": false,
						"schema": { "type": "string", "pattern": "^\\d+$" }
					},
					{
						"name": "operationType",
						"in": "query",
						"required": false,
						"schema": { "type": "string" }
					},
					{
						"name": "from",
						"in": "query",
						"required": false,
						"schema": { "type": "string", "format": "date-time" }
					},
					{
						"name": "to",
						"in": "query",
						"required": false,
						"schema": { "type": "string", "format": "date-time" }
					}
				],
				"responses": {
					"200": {
						"description": "Operation search result",
						"content": {
							"application/json": {
								"schema": { "$ref": "#/components/schemas/ExplorerOperations" }
							}
						}
					},
					"400": { "description": "Invalid operation filters" },
					"502": { "description": "Operation search unavailable" }
				}
			}
		},
		"/v1/explorer/contracts/{contractId}": {
			"get": {
				"summary": "Look up contract readiness",
				"description": "Reports RPC-backed contract lookup readiness for a contract ID.",
				"tags": ["Explorer", "Horizon"],
				"operationId": "getExplorerContract",
				"parameters": [
					{
						"name": "contractId",
						"in": "path",
						"required": true,
						"schema": { "type": "string", "pattern": "^C[A-Z2-7]{55}$" }
					}
				],
				"responses": {
					"200": {
						"description": "Contract lookup readiness",
						"content": {
							"application/json": {
								"schema": { "$ref": "#/components/schemas/ExplorerContract" }
							}
						}
					},
					"400": { "description": "Invalid contract id" }
				}
			}
		},
...
		"/v1/transactions/{hash}": {
			"get": {
				"summary": "Look up a transaction by hash",
				"description": "Returns lean public transaction facts from the configured Horizon source. This is a lookup boundary, not the full-history graph API.",
				"tags": ["Explorer", "Horizon"],
				"operationId": "getTransactionByHash",
				"parameters": [
					{
						"name": "hash",
						"in": "path",
						"description": "64-character transaction hash",
						"required": true,
						"schema": {
							"type": "string",
							"pattern": "^[0-9a-fA-F]{64}$"
						}
					}
				],
				"responses": {
					"200": {
						"description": "Transaction lookup result",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/TransactionLookup"
								}
							}
						}
					},
					"400": {
						"description": "Invalid transaction hash",
						"content": {
							"application/json": {
								"schema": { "$ref": "#/components/schemas/ErrorResponse" }
							}
						}
					},
					"404": {
						"description": "Transaction not found in configured Horizon",
						"content": {
							"application/json": {
								"schema": { "$ref": "#/components/schemas/ErrorResponse" }
							}
						}
					},
					"502": {
						"description": "Transaction lookup unavailable",
						"content": {
							"application/json": {
								"schema": { "$ref": "#/components/schemas/ErrorResponse" }
							}
						}
					}
				}
			}
		},
...
	},
	"components": {
		"schemas": {
...
			"ErrorResponse": {
				"type": "object",
				"required": ["error"],
				"properties": {
					"error": { "type": "string" }
				}
			},
			"TransactionLookup": {
				"type": "object",
				"required": [
					"createdAt",
					"feeCharged",
					"hash",
					"ledger",
					"operationCount",
					"source",
					"sourceAccount",
					"successful"
				],
				"properties": {
					"createdAt": {
						"type": "string",
						"format": "date-time"
					},
					"feeCharged": { "type": "string" },
					"hash": { "type": "string" },
					"ledger": { "type": "string" },
					"operationCount": { "type": "integer", "minimum": 0 },
					"source": {
						"type": "string",
						"enum": ["horizon"]
					},
					"sourceAccount": { "type": "string" },
					"successful": { "type": "boolean" }
				}
			},
...
			"RecentTransactions": {
				"type": "object",
				"required": ["generatedAt", "limit", "records", "source", "truncated"],
				"properties": {
					"generatedAt": { "type": "string", "format": "date-time" },
					"limit": { "type": "integer", "minimum": 1, "maximum": 50 },
					"records": {
						"type": "array",
						"items": { "$ref": "#/components/schemas/TransactionLookup" }
					},
					"source": { "type": "string", "enum": ["horizon"] },
					"truncated": { "type": "boolean" }
				}
			},
			"ExplorerLedger": {
				"type": "object",
				"required": [
					"closedAt",
					"hash",
					"operationCount",
					"protocolVersion",
					"sequence",
					"source",
					"transactionCount"
				],
				"properties": {
					"closedAt": { "type": "string", "format": "date-time" },
					"hash": { "type": "string" },
					"operationCount": { "type": "integer", "minimum": 0 },
					"protocolVersion": { "type": "integer" },
					"sequence": { "type": "string" },
					"source": { "type": "string", "enum": ["horizon"] },
					"transactionCount": {
						"type": "integer",
						"nullable": true,
						"minimum": 0
					}
				}
			},
			"ExplorerAccountBalance": {
				"type": "object",
				"required": ["assetCode", "assetIssuer", "assetType", "balance"],
				"properties": {
					"assetCode": { "type": "string", "nullable": true },
					"assetIssuer": { "type": "string", "nullable": true },
					"assetType": { "type": "string" },
					"balance": { "type": "string" }
				}
			},
			"ExplorerAccount": {
				"type": "object",
				"required": [
					"accountId",
					"balances",
					"lastModifiedLedger",
					"sequence",
					"source",
					"subentryCount"
				],
				"properties": {
					"accountId": { "type": "string" },
					"balances": {
						"type": "array",
						"items": { "$ref": "#/components/schemas/ExplorerAccountBalance" }
					},
					"lastModifiedLedger": { "type": "string", "nullable": true },
					"sequence": { "type": "string" },
					"source": { "type": "string", "enum": ["horizon"] },
					"subentryCount": { "type": "integer", "minimum": 0 }
				}
			},
			"ExplorerAsset": {
				"type": "object",
				"required": [
					"amount",
					"assetCode",
					"assetIssuer",
					"assetType",
					"numAccounts",
					"source"
				],
				"properties": {
					"amount": { "type": "string", "nullable": true },
					"assetCode": { "type": "string", "nullable": true },
					"assetIssuer": { "type": "string", "nullable": true },
					"assetType": { "type": "string" },
					"numAccounts": {
						"type": "integer",
						"nullable": true,
						"minimum": 0
					},
					"source": { "type": "string", "enum": ["horizon"] }
				}
			},
			"ExplorerAssets": {
				"type": "object",
				"required": ["assets", "source", "truncated"],
				"properties": {
					"assets": {
						"type": "array",
						"items": { "$ref": "#/components/schemas/ExplorerAsset" }
					},
					"source": { "type": "string", "enum": ["horizon"] },
					"truncated": { "type": "boolean" }
				}
			},
			"ExplorerOperationFilters": {
				"type": "object",
				"properties": {
					"accountId": { "type": "string" },
					"from": { "type": "string", "format": "date-time" },
					"ledger": { "type": "string" },
					"operationType": { "type": "string" },
					"to": { "type": "string", "format": "date-time" },
					"transactionHash": { "type": "string" }
				}
			},
			"ExplorerOperation": {
				"type": "object",
				"required": [
					"createdAt",
					"id",
					"ledger",
					"source",
					"sourceAccount",
					"successful",
					"transactionHash",
					"type",
					"typeNumber"
				],
				"properties": {
					"createdAt": { "type": "string", "format": "date-time" },
					"id": { "type": "string" },
					"ledger": { "type": "string", "nullable": true },
					"source": { "type": "string", "enum": ["horizon"] },
					"sourceAccount": { "type": "string", "nullable": true },
					"successful": { "type": "boolean", "nullable": true },
					"transactionHash": { "type": "string", "nullable": true },
					"type": { "type": "string" },
					"typeNumber": { "type": "integer", "nullable": true }
				}
			},
			"ExplorerOperations": {
				"type": "object",
				"required": ["filters", "records", "source", "truncated"],
				"properties": {
					"filters": {
						"$ref": "#/components/schemas/ExplorerOperationFilters"
					},
					"records": {
						"type": "array",
						"items": { "$ref": "#/components/schemas/ExplorerOperation" }
					},
					"source": { "type": "string", "enum": ["horizon"] },
					"truncated": { "type": "boolean" }
				}
			},
			"ExplorerContract": {
				"type": "object",
				"required": ["contractId", "message", "source", "status"],
				"properties": {
					"contractId": { "type": "string" },
					"message": { "type": "string" },
					"source": { "type": "string", "enum": ["rpc"] },
					"status": {
						"type": "string",
						"enum": ["loaded", "unavailable", "unconfigured"]
					}
				}
			},
			"ExplorerSearch": {
				"type": "object",
				"required": ["query", "result", "resultType", "source"],
				"properties": {
					"query": { "type": "string" },
					"result": {
						"nullable": true,
						"oneOf": [
							{ "$ref": "#/components/schemas/ExplorerAccount" },
							{ "$ref": "#/components/schemas/ExplorerAssets" },
							{ "$ref": "#/components/schemas/ExplorerContract" },
							{ "$ref": "#/components/schemas/ExplorerLedger" },
							{ "$ref": "#/components/schemas/ExplorerOperation" },
							{ "$ref": "#/components/schemas/TransactionLookup" }
						]
					},
					"resultType": {
						"type": "string",
						"enum": [
							"account",
							"asset",
							"contract",
							"ledger",
							"not_found",
							"transaction",
							"unknown"
						]
					},
					"source": { "type": "string", "enum": ["horizon", "rpc"] }
				}
			},
...
			"SearchFacetValue": {
				"type": "object",
				"required": ["count", "value"],
				"properties": {
					"count": { "type": "integer" },
					"value": { "type": "string" }
				}
			},
			"SearchFacets": {
				"type": "object",
				"required": [
					"active",
					"archiveStatus",
					"countryCode",
					"entityType",
					"fullValidator",
					"topTier",
					"validating",
					"validator"
				],
				"properties": {
					"active": {
						"type": "array",
						"items": { "$ref": "#/components/schemas/SearchFacetValue" }
					},
					"archiveStatus": {
						"type": "array",
						"items": { "$ref": "#/components/schemas/SearchFacetValue" }
					},
					"countryCode": {
						"type": "array",
						"items": { "$ref": "#/components/schemas/SearchFacetValue" }
					},
					"entityType": {
						"type": "array",
						"items": { "$ref": "#/components/schemas/SearchFacetValue" }
					},
					"fullValidator": {
						"type": "array",
						"items": { "$ref": "#/components/schemas/SearchFacetValue" }
					},
					"topTier": {
						"type": "array",
						"items": { "$ref": "#/components/schemas/SearchFacetValue" }
					},
					"validating": {
						"type": "array",
						"items": { "$ref": "#/components/schemas/SearchFacetValue" }
					},
					"validator": {
						"type": "array",
						"items": { "$ref": "#/components/schemas/SearchFacetValue" }
					}
				}
			},
			"SearchHit": {
				"type": "object",
				"required": ["detail", "entityId", "entityType", "href", "id", "label"],
				"properties": {
					"detail": { "type": "string" },
					"entityId": { "type": "string" },
					"entityType": {
						"type": "string",
						"enum": ["node", "organization"]
					},
					"href": { "type": "string" },
					"id": { "type": "string" },
					"label": { "type": "string" },
					"organizationName": { "type": "string" }
				}
			},
			"SearchReadModel": {
				"type": "object",
				"required": ["fallbackReason", "schemaVersion"],
				"properties": {
					"fallbackReason": {
						"type": "string",
						"nullable": true,
						"enum": [
							"meilisearch_syncing",
							"meilisearch_unavailable",
							"meilisearch_unconfigured"
						]
					},
					"schemaVersion": { "type": "string", "enum": ["v1"] }
				}
			},
			"SearchResponse": {
				"type": "object",
				"required": [
					"estimatedTotalHits",
					"facets",
					"hits",
					"indexedNetworkTime",
					"query",
					"readModel",
					"source"
				],
				"properties": {
					"estimatedTotalHits": { "type": "integer" },
					"facets": { "$ref": "#/components/schemas/SearchFacets" },
					"hits": {
						"type": "array",
						"items": { "$ref": "#/components/schemas/SearchHit" }
					},
					"indexedNetworkTime": { "type": "string" },
					"query": { "type": "string" },
					"readModel": { "$ref": "#/components/schemas/SearchReadModel" },
					"source": {
						"type": "string",
						"enum": ["memory", "meilisearch"]
					}
				}
			}
		}
	}
}